+++ lib/public/utilities/fetch/RemoteDataSource.js
@@ -68,62 +68,3 @@ export class RemoteDataSource extends Observable {
         return this._observableData;
     }
 }
-
-// eslint-disable-next-line valid-jsdoc
-/**
- * @inheritDoc
- */
-export class PaginatableRemoteDataSource extends RemoteDataSource {
-    /**
-     * Constructor
-     * @param {object} [options] options modifying behaviour
-     * @param {Function<U, V>} [options.dataTransformer = null] function to transform response of server
-     * it is called if keepExisting flag is specified @see PaginatableRemoteDataSource.acquire
-     * @param {Function<U, void>} [options.onSuccess = null] function called with raw response from server if successful
-     */
-    constructor(options) {
-        super();
-        this._options = options;
-        this._endpoint = null;
-        this._observablePage = new ObservableData();
-    }
-
-    /**
-     * Fetch the given endpoint to fill current data
-     *
-     * @param {string} endpoint the endpoint to fetch
-     * @return {Promise<void>} resolves once the data fetching has ended
-     */
-    async fetch(endpoint, { keepExisting }) {
-        const abortController = new AbortController();
-        try {
-            if (this._abortController && !keepExisting) {
-                this._abortController.abort();
-                this._observableData.current = RemoteData.loading();
-            }
-            this._abortController = abortController;
-            const { data, meta: { page } = {} } = await getRemoteData(endpoint, { signal: this._abortController.signal });
-            this._observableData.current = RemoteData.success([
-                ... keepExisting ? this._observableData.current.payload ?? [] : [],
-                ... this._options.dataTransformer?.(data) ?? data,
-            ]);
-            const { totalCount = data.length, pageCount } = page ?? {};
-            this._observablePage.current = { totalCount, pageCount };
-        } catch (error) {
-            // Use local variable because the class member (this._abortController) may already have been override in another call
-            if (!abortController.signal.aborted) {
-                this._observableData.current = RemoteData.failure(error);
-            }
-            throw error;
-        }
-    }
-
-    /**
-     * Return the observable remote data provided by this source
-     *
-     * @return {ObservableData<{totalCount: number, pageCount: number}>} the observable remote data
-     */
-    get observablePage() {
-        return this._observablePage;
-    }
-}
