+++ lib/public/views/lhcPeriods/Overview/LhcPeriodsOverviewModel.js
@@ -16,7 +16,7 @@ import { PaginationModel } from '../../../components/Pagination/PaginationModel.
 import { SortModel } from '../../../components/common/table/SortModel.js';
 import { TextTokensFilterModel } from '../../../components/Filters/common/filters/TextTokensFilterModel.js';
 import { buildUrl } from '../../../utilities/fetch/buildUrl.js';
-import { PaginatableRemoteDataSource } from '../../../utilities/fetch/RemoteDataSource.js';
+import { PaginatableRemoteDataSource } from '../../../utilities/fetch/PaginatableRemoteDataSource.js';
 
 /**
  * LHC Periods overview model
